refactor: added new component IngredientPicture.tsx
