+++ components/AppBar.tsx
@@ -1,24 +1,13 @@
 "use client";
 
 import * as React from "react";
-import { useSearchParams } from "next/navigation";
-import IngredientSelector from "@/components/IngredientSelector";
-
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import { selectCurrentIngredients, selectURLParams } from "@/redux/selectors";
 import store, { RootState } from "@/redux/store";
-import IngredientCard from "@/components/IngredientCard";
-import { Button, Icon, Snackbar, Stack } from "@mui/material";
-import { openEditor } from "@/redux/reducers/editor";
-import { updateRecipeIngredients } from "@/redux/reducers/recipes";
-import ShowNutritionalTable from "@/components/ShowNutritionalTable";
-
+import { Snackbar } from "@mui/material";
 import Box from "@mui/joy/Box";
-
 import ShareIcon from "@mui/icons-material/Share";
 import SaveIcon from "@mui/icons-material/Save";
-import Typography from "@mui/joy/Typography";
-import Grid from "@mui/joy/Grid";
 import Sheet from "@mui/joy/Sheet";
 import { IconButton, Tooltip } from "@mui/joy";
