+++ components/IngredientCards/IngredientPicture.tsx
